Apply LoRas to refiner
The stage for applying LoRas currently only applies selected LoRas to
the base model. The more steps run using a refiner model, the less the
effects of the selected LoRas can be seen in the output image because
the refiner model never has the LoRas applied to it. This change makes
it so that, when a refiner model is selected, the LoRas are applied to
that along with the base model.